V1/portfolio sandeep

+++ backend/routes/posts.js
@@ -26,14 +26,16 @@ const storage= multer.diskStorage({
   }
 })
 
-router.post("",
+router.post("",checkAuth,
 multer({storage: storage}).single("image"),(req, res, next)=>{
   const url= req.protocol + '://' + req.get("host");
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
+
   post.save().then(createdPost=>{
 
     res.status(201).json({
@@ -86,7 +88,8 @@ router.get('',(req,res, next)=> {
 
 
 
-router.put("/:id",checkAuth,multer({storage: storage}).single("image"),
+router.put("/:id",
+checkAuth,multer({storage: storage}).single("image"),
 (req, res, next)=>{
 
   let imagePath=req.body.imagPath;
@@ -102,14 +105,17 @@ router.put("/:id",checkAuth,multer({storage: storage}).single("image"),
     title: req.body.title,
     content: req.body.content,
     imagePath:imagePath,
-    image:req.body.image
+    image:req.body.image,
+    creator:req.userData.userId
   };
 try {
-  Post.updateOne({_id: req.params.id},{$set:post}).then(result=>{
-
-
-    res.status(200).json({message: 'Update successful'});
-
+  Post.updateOne({_id: req.params.id,creator: req.userData.userId},{$set:post}).then(result=>{
+    if(result.modifiedCount > 0){
+      res.status(200).json({message:"Update  successful"})
+    }
+    else{
+    res.status(401).json({message: 'Not Authorized!'});
+    }
   });
 } catch (error) {
   console.log(error);
@@ -133,9 +139,14 @@ router.get("/:id", (req, res, next)=>
 
 router.delete("/:id",checkAuth,
  (req, res, next) => {
- Post.deleteOne({_id: req.params.id}).then(result =>{
-  console.log(result);
-  res.status(200).json({ message: "post deleted!"});
+ Post.deleteOne({_id: req.params.id,creator:req.userData.userId}).then(result =>{
+  if(result.deletedCount > 0){
+    res.status(200).json({ message: "post deleted!"});
+  }
+  else{
+  res.status(401).json({message: 'Not Authorized!'});
+  }
+
  });
 });
